ms_calculations/mutagen/requirements/req_3_10_13.txt
anyio==4.3.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
duckdb==0.9.2
exceptiongroup==1.2.0
filelock==3.13.1
fontawesomefree==6.5.1
fonttools==4.45.1
fsspec==2023.10.0
h11==0.14.0
-e git+https://github.com/tconbeer/harlequin.git@a7618dee44501c91c43dd73182ab070a0906e8d7#egg=harlequin
httpcore==1.0.4
httpx==0.27.0
idna==3.6
Jinja2==3.1.2
kiwisolver==1.4.5
linkify-it-py==2.0.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdit-py-plugins==0.4.0
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==2.1.4
patsy==0.5.4
Pillow==10.1.0
platformdirs==4.1.0
plotly==5.18.0
prompt-toolkit==3.0.36
pyarrow==14.0.2
pyasn1==0.5.1
pycparser==2.21
Pygments==2.17.2
pyOpenSSL==24.0.0
pyparsing==3.1.1
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2023.3.post1
pywaffle==1.1.0
questionary==2.0.1
requests==2.31.0
rich==13.7.0
rich-click==1.7.3
scipy==1.11.4
seaborn==0.13.2
shandy-sqlfmt==0.21.1
six==1.16.0
sniffio==1.3.1
squarify==0.4.3
statsmodels==0.14.1
sympy==1.12
tenacity==8.2.3
textual==0.46.0
textual-fastdatatable==0.6.3
textual-textarea==0.9.5
tomli==2.0.1
tomlkit==0.12.3
torch==2.1.1
torchaudio==2.1.1
torchinfo==1.8.0
torchvision==0.16.1
tqdm==4.66.1
tree-sitter==0.20.4
tree-sitter-languages==1.9.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.1.0
wcwidth==0.2.13
openpyxl